Refactor of prover and verifier functions
