+++ proving-system/stark/src/utils.rs
@@ -1,0 +1,60 @@
+use super::{constraints::boundary::BoundaryConstraints, FE, ORDER_OF_ROOTS_OF_UNITY_TRACE};
+use lambdaworks_math::polynomial::Polynomial;
+use std::ops::Div;
+
+pub fn fibonacci_trace(initial_values: [FE; 2]) -> Vec<FE> {
+    let mut ret: Vec<FE> = vec![];
+
+    ret.push(initial_values[0].clone());
+    ret.push(initial_values[1].clone());
+
+    for i in 2..(ORDER_OF_ROOTS_OF_UNITY_TRACE) {
+        ret.push(ret[i - 1].clone() + ret[i - 2].clone());
+    }
+
+    ret
+}
+
+pub(crate) fn compute_boundary_quotient(
+    constraints: &BoundaryConstraints<FE>,
+    col: usize,
+    primitive_root: &FE,
+    trace_poly: &Polynomial<FE>,
+) -> Polynomial<FE> {
+    let domain = constraints.generate_roots_of_unity(primitive_root);
+    let values = constraints.values(col);
+    let zerofier = constraints.compute_zerofier(primitive_root);
+
+    let poly = Polynomial::interpolate(&domain, &values);
+
+    (trace_poly.clone() - poly).div(zerofier)
+}
+
+pub(crate) fn compute_zerofier(primitive_root: &FE, root_order: usize) -> Polynomial<FE> {
+    let roots_of_unity_vanishing_polynomial =
+        Polynomial::new_monomial(FE::one(), root_order) - Polynomial::new(&[FE::one()]);
+    let exceptions_to_vanishing_polynomial =
+        Polynomial::new(&[-primitive_root.pow(root_order - 2), FE::one()])
+            * Polynomial::new(&[-primitive_root.pow(root_order - 1), FE::one()]);
+
+    roots_of_unity_vanishing_polynomial.div(exceptions_to_vanishing_polynomial)
+}
+
+/// This functions takes a roots of unity and a coset factor
+/// If coset_factor is 1, it's just expanding the roots of unity
+/// w ^ 0, w ^ 1, w ^ 2 .... w ^ n-1
+/// If coset_factor is h
+/// h * w ^ 0, h * w ^ 1 .... h * w ^ n-1
+pub(crate) fn generate_roots_of_unity_coset(coset_factor: u64, primitive_root: &FE) -> Vec<FE> {
+    let coset_factor: FE = coset_factor.into();
+
+    let mut numbers = vec![coset_factor.clone()];
+    let mut exp: u64 = 1;
+    let mut next_root = primitive_root.pow(exp) * &coset_factor;
+    while next_root != coset_factor {
+        numbers.push(next_root);
+        exp += 1;
+        next_root = primitive_root.pow(exp) * &coset_factor;
+    }
+    numbers
+}
